Refactor Kotlin Inject components by scope
