+++ app/src/main/java/com/archstarter/app/MyApp.kt
@@ -1,4 +1,7 @@
 package com.archstarter.app
 
 import android.app.Application
-class MyApp : Application()
+
+class MyApp : Application() {
+    val appScopeManager: AppScopeManager by lazy { AppScopeManager() }
+}
